Fixing frequency of heartbeat messages in UAVs
Two things were wrong here. First, a random delay is only necessary on the first timer. It's necessary to prevent message collisions but after the first random delay every subsequent one is also guaranteed not to colide.

The other thing that was wrong is that there was no fixed delay, only the random one. This made it so heartbeat messages are sent way too frequently.